Renamed the gem to logging_library
Also added an explicit MIT license, since this library is now released as [FOSS](https://en.wikipedia.org/wiki/Free_and_open-source_software)
