+++ README.md
@@ -65,7 +65,7 @@ push git commits and tags, and push the `.gem` file to
 
 ## Contributing
 
-Bug reports and pull requests are welcome on GitHub at 
+Bug reports and pull requests are welcome on GitHub at
 https://github.com/ecraft/ecraft-logging_library. This project is intended to
 be a safe, welcoming space for collaboration, and contributors are expected to
 adhere to the [Contributor Covenant](http://contributor-covenant.org) code of
@@ -83,3 +83,6 @@ Preview API documentation locally with
 bundle exec yard server -r
 ```
 
+### License
+
+MIT
