+++ lib/ecraft/logging_library.rb
@@ -1,11 +1,4 @@
-require 'ecraft/logging_library/custom_formatter'
-require 'ecraft/logging_library/logger_factory'
-require 'ecraft/logging_library/logger'
-require 'ecraft/logging_library/mixins/loggable'
-require 'ecraft/logging_library/version'
-
 module Ecraft
-  module LoggingLibrary
-    Loggable = Mixins::Loggable
-  end
+  # For backwards compatibility, since this library used to be called ecraft-logging_library before it was released as FOSS.
+  LoggingLibrary = ::LoggingLibrary
 end
