--- lib/ecraft/logging_library/version.rb
@@ -1,5 +1,0 @@
-module Ecraft
-  module LoggingLibrary
-    VERSION = '1.0.3'.freeze
-  end
-end
