+++ lib/logging_library/version.rb
@@ -1,0 +1,3 @@
+module LoggingLibrary
+  VERSION = '1.0.3'.freeze
+end
