+++ spec/logging_library_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+
+describe LoggingLibrary do
+  it 'has a version number' do
+    expect(LoggingLibrary::VERSION).not_to be nil
+  end
+end
